estrutura modular, inicio, funcional
