+++ index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <link rel="icon" type="image/png" href="./favicon.ico" />
+    <link rel="stylesheet" type="text/css" href="./style.css" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>IFC.js</title>
+  </head>
+  <body>
+    <div id="input-container">
+      <input readonly type="file" id="openFileDialog" />
+    </div>
+    <canvas id="threeCanvas"></canvas>\
+    <p style="padding-left: 150px;">Hello World</p>
+    <a id ="test-button" class="button4" style="background-color:#a79e1cef">TEST</a>
+    <script type="module" src="bundle.js"></script>
+  </body>
+</html>
