+++ style.css
@@ -1,0 +1,36 @@
+* {
+    padding: 0;
+    margin: 0;
+}
+
+html, body {
+    overflow: hidden;
+}
+
+#threeCanvas {
+    position: fixed;
+    top: 0;
+    left: 0;
+    outline: none;
+}
+
+#input-container {
+    z-index: 1;
+    position: fixed;
+    left: 7rem;
+    display: flex;
+    flex-direction: column;
+    top: 1rem;
+}
+
+#input-container input {
+   margin-top: 1rem;
+}
+
+#close-button {
+    z-index: 1;
+    position: fixed;
+    top: 3rem;
+    padding: 0 2.9rem;
+    left: 7rem;
+}
